+++ src/lib/jwk-store.ts
@@ -18,39 +18,12 @@
  * @module lib/jwk-store
  */
 
-import { randomBytes } from 'node:crypto';
-
 import type { GenerateKeyPairOptions } from 'jose';
 import { exportJWK, importJWK, generateKeyPair } from 'jose';
 
 import type { JWK } from './types';
-import type { JWKWithKid } from './types-internals';
-import {
-  assertIsPlainObject,
-  privateToPublicKeyTransformer,
-  supportedAlgs,
-} from './helpers';
-
-const generateRandomKid = () => {
-  return randomBytes(40).toString('hex');
-};
-
-function normalizeKeyKid(
-  jwk: unknown,
-  opts?: { kid?: string },
-): asserts jwk is JWKWithKid {
-  assertIsPlainObject(jwk, 'Invalid jwk format');
-
-  if (jwk['kid'] !== undefined) {
-    return;
-  }
-
-  if (opts?.kid !== undefined) {
-    jwk['kid'] = opts.kid;
-  } else {
-    jwk['kid'] = generateRandomKid();
-  }
-}
+import { privateToPublicKeyTransformer, supportedAlgs } from './helpers';
+import { assertIsJwtWithKid } from './assertions';
 
 /**
  * Simple JWK store
@@ -94,7 +67,7 @@ export class JWKStore {
 
     const pair = await generateKeyPair(alg, generateOpts);
     const jwk = await exportJWK(pair.privateKey);
-    normalizeKeyKid(jwk, opts);
+    assertIsJwtWithKid(jwk, opts);
     jwk.alg = alg;
 
     this.#keyRotator.add(jwk);
@@ -109,7 +82,7 @@ export class JWKStore {
   async add(maybeJwk: Record<string, unknown>): Promise<JWK> {
     const jwk = { ...maybeJwk };
 
-    normalizeKeyKid(jwk);
+    assertIsJwtWithKid(jwk);
 
     if (!('alg' in jwk)) {
       throw new Error('Unspecified JWK "alg" property');
